+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - double quotes in a host now raises an error
 - calling `require 'addressable'` will now load both the URI and Template files
 - assigning to the `hostname` component with an `IPAddr` object is now supported
+- assigning to the `origin` component is now supported
 
 # Addressable 2.3.8
 - fix warnings